requirements.txt
ccxt>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
ta>=0.11.0
PyYAML>=6.0
APScheduler>=3.10.0
python-telegram-bot>=20.0
python-dotenv>=1.0.0
